Custom summary bandwidth stats charts.
Meant to replace the vnstati generated summary chart. Provides
information on today, yesterday, this month, last month, and total
bandwidth usage.

Note that no precautions were taken for the case where vnstat is newly
installed and so has no data for yesterday/last month. I didn't want to
wipe my stats to test it. Will revisit later when I move to a new
seedbox.

I notice that the daily stats from vnstati were calculated based on the
time that the computer was actually on. Can't figure out how they got
that from the data available in dumpdb, so maybe it's some extra data
they store in the database that isn't available?

The daily average is calculated assuming 24 hour days.

The monthly average is calculated based on elapsed seconds since the
monthly entry's creation timestamp. e.g.: If vnstat started recording
from on the 15th of last month, it'll only count the secconds between
the 15th and the last day-hour-second of the month. This seems to be
different from vnstati calculates averages, it seems that vnstati
calculates average for the entire month.

I've figured out what the total rx/tx KiB counters do. Since the data
doesn't record in floats, only in ints, they count the leftover bytes
that didn't get included in the MiB counter.

Using AwesomeChart to generate the chart graphics in javascript.
Downside is that because it's using html5's canvas, you have to state
fixed width & height for the canvas. Also had to hardcode colours in
javascript for the charts.